requirements.txt
asgiref==3.7.2
Django==4.2.5
Pillow==10.0.1
psycopg2-binary==2.9.8
sqlparse==0.4.4
djangorestframework==3.14.0
django-filter==23.3